+++ lib/domino_bulma.dart
@@ -3,4 +3,4 @@ export 'src/_components.dart';
 export 'src/_elements.dart';
 export 'src/_form.dart';
 export 'src/_layout.dart';
-export 'src/_modifiers.dart' show TypeMod, SizeMod, FloatMod;
+export 'src/_modifiers.dart' show MainColor, SizeMod, FloatMod;